feat: move models to separate package to fix cycle imports
